Navbar: move control buttons to dedicated component
