+++ src/components/navbar/Navbar.svelte
@@ -1,14 +1,10 @@
 <script lang="ts">
-  import { dev } from '$app/environment';
-  import { goto } from '$app/navigation';
-  import api from '$endpoints';
-  import Separator from '$components/Separator.svelte';
   import type { ItemId } from '$lib/itemId';
   import type { ItemData } from '$lib/server/data/item/item';
   import { getDescendant } from '$lib/itemData';
   import consts from '$lib/consts';
-  import { Button } from '$components/base';
   import { onMount } from 'svelte';
+  import ControlButtons from './ControlButtons.svelte';
 
   type NavbarPath = { url: string; txt: string }[];
 
@@ -96,25 +92,6 @@
     }
   });
 
-  /** Log out, then reload the page */
-  async function logOut() {
-    await api().admin.auth.logout();
-    location.reload();
-  }
-
-  /**
-   * Go to the login page, with the from parameter determining the origin page.
-   */
-  async function gotoLogin() {
-    await goto(`/admin/login?from=${window.location.pathname}`);
-  }
-
-  /** Clear all data, and take the user to the firstrun page */
-  async function clear() {
-    await api().debug.clear();
-    await goto('/admin/firstrun/account');
-  }
-
   // This function needs to accept `path` as an input, otherwise the links
   // stop being reactive due to caching or something
   function pathTo(path: NavbarPath, i: number) {
@@ -129,14 +106,6 @@
     // https://css-tricks.com/styling-based-on-scroll-position/
     document.documentElement.dataset.scroll =
       window.scrollY < 20 ? 'top' : 'page';
-  }
-
-  function editButtonClick() {
-    if (editing) {
-      onEditFinish?.();
-    } else {
-      onEditBegin?.();
-    }
   }
 
   // Update scroll position on mount so that we don't get weird visual artifacts
@@ -157,32 +126,13 @@
     </h1>
   </span>
 
-  <!-- Control buttons -->
-  <span id="control-buttons">
-    {#if loggedIn}
-      {#if editable}
-        <!-- Use the one button so that animations persist between states -->
-        <Button
-          onclick={editButtonClick}
-          mode={editing ? 'confirm' : 'default'}
-        >
-          {editing ? 'Finish editing' : 'Edit'}
-        </Button>
-      {/if}
-      <Button onclick={() => goto('/admin')}>Admin</Button>
-      <Button onclick={logOut}>Log out</Button>
-    {:else if loggedIn !== undefined}
-      <!-- Only include a login button if logging in is enabled -->
-      <Button onclick={gotoLogin}>Log in</Button>
-    {/if}
-    <!-- About button navigates to about page -->
-    <Button onclick={() => goto('/about')}>About</Button>
-    <!-- In dev mode, add a quick shortcut to delete everything -->
-    {#if dev}
-      <Separator />
-      <Button onclick={clear} mode="warning">Clear data</Button>
-    {/if}
-  </span>
+  <ControlButtons
+    {loggedIn}
+    {editable}
+    {editing}
+    {onEditBegin}
+    {onEditFinish}
+  />
 </nav>
 
 <style>
@@ -217,14 +167,5 @@
   h1 {
     margin: 0px;
     font-size: 3em;
-  }
-
-  #control-buttons {
-    display: flex;
-    gap: 5px;
-    align-items: center;
-    justify-content: center;
-    grid-area: control-buttons;
-    margin-right: 20px;
   }
 </style>
